@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 android {
-        compileSdkVersion 27
+        compileSdkVersion 28
                 defaultConfig {
                         applicationId "com.weihuagu.receiptnotice"
                                 minSdkVersion 19
-                                targetSdkVersion 25
-                                versionCode 12
-                                versionName "1.1"
-                                testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+                                targetSdkVersion 28
+                                versionCode 16
+                                versionName "1.5"
+                                testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
                 }
 
 
@@ -32,20 +32,22 @@ android {
 }
 dependencies {
                 implementation fileTree(dir: 'libs', include: ['*.jar'])
-                implementation 'com.android.support:appcompat-v7:27.0.2'
-                implementation 'com.android.support:design:27.0.2'
-                implementation 'com.android.support:cardview-v7:27.0.2'
-                implementation 'com.android.support:recyclerview-v7:27.0.2'
-                implementation 'com.android.support.constraint:constraint-layout:1.0.2'
+                implementation 'androidx.appcompat:appcompat:1.0.0'
+                implementation 'com.google.android.material:material:1.0.0'
+                implementation 'androidx.cardview:cardview:1.0.0'
+                implementation 'androidx.recyclerview:recyclerview:1.0.0'
+                implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+                implementation "androidx.viewpager2:viewpager2:1.0.0-beta04"
                 implementation 'com.squareup.okhttp3:okhttp:3.12.1'
                 implementation 'com.github.wangjintao:TLog:v1.0.1'
                 implementation 'com.github.pedrovgs:lynx:1.1.0'
                 implementation 'io.socket:socket.io-client:1.0.0'
                 implementation 'com.google.code.gson:gson:2.8.5'
+                implementation 'com.jeremyliao:live-event-bus-x:1.5.7'
                 testImplementation 'junit:junit:4.12'
                 testImplementation "org.mockito:mockito-core:3.+"
-                androidTestImplementation 'com.android.support.test:runner:1.0.1'
-                androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
+                androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+                androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
                 androidTestImplementation 'org.mockito:mockito-android:+'
 }
 
