@@ -2,8 +2,6 @@ package com.weihuagu.receiptnotice;;
 
 import android.app.ActivityManager;
 import android.app.Service;
-import android.app.AlarmManager;
-import android.app.PendingIntent;
 import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
@@ -16,33 +14,36 @@ import android.os.PowerManager.WakeLock;
 import android.os.PowerManager;
 import io.socket.client.IO;
 import io.socket.client.Socket;
+
+import java.lang.reflect.Field;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Date;
 import java.util.ArrayList;
-import java.util.List;
+import java.util.Map;
 import java.util.Random;
-import java.util.concurrent.TimeUnit;
 import java.lang.System;
 import java.lang.Thread;
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.net.URISyntaxException;
 import java.text.SimpleDateFormat;
-import java.io.IOException;
 import java.security.KeyManagementException;
 import java.security.NoSuchAlgorithmException;
 import javax.net.ssl.SSLSocketFactory;
 import com.google.gson.Gson;
+import com.weihuagu.receiptnotice.util.DeviceInfoUtil;
+import com.weihuagu.receiptnotice.util.ExternalInfoUtil;
+import com.weihuagu.receiptnotice.util.LogUtil;
+import com.weihuagu.receiptnotice.util.PreferenceUtil;
+import com.weihuagu.receiptnotice.util.SSLSocketFactoryCompat;
+
 import io.socket.emitter.Emitter;
-import org.json.JSONException;
-import org.json.JSONObject;
+
 import java.util.Timer;
 import java.util.TimerTask;
-import okhttp3.MediaType;
+
 import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.RequestBody;
-import okhttp3.Response;
 import okhttp3.TlsVersion;
 import okhttp3.ConnectionSpec;
 
@@ -78,7 +79,7 @@ public class NotificationCollectorMonitorService extends Service {
                 PowerManager pm = (PowerManager)getSystemService(
                                 Context.POWER_SERVICE);
                 wl = pm.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK,
-                                "receiptnoticewakelock");
+                                "receiptnotice:NotificationCollectorMonitorServicewakelock");
                 wl.acquire();
 
         }
@@ -177,18 +178,34 @@ public class NotificationCollectorMonitorService extends Service {
                 PreferenceUtil preference=new PreferenceUtil(getBaseContext());
                 Gson gson = new Gson();
                 if(preference. isEcho()&&(preference.getEchoServer()!=null)){
-                        Date date=new Date(System.currentTimeMillis());
-                        SimpleDateFormat format=new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-                        String time=format.format(date);
-                        DeviceBean device=new DeviceBean();
-                        String deviceid=preference.getDeviceid();
-                        deviceid=(!deviceid.equals("") ? deviceid:DeviceInfoUtil.getUniquePsuedoID());
-                        device.setDeviceid(deviceid);
-                        device.setTime(time);
-                        LogUtil.debugLog("start connect socketio");
-                        echoServerBySocketio(preference.getEchoServer(),gson.toJson(device));
-                        LogUtil.debugLog(gson.toJson(device));
-                        return true;
+
+                                Date date = new Date(System.currentTimeMillis());
+                                SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+                                String time = format.format(date);
+                                DeviceBean device = new DeviceBean();
+                                String deviceid = preference.getDeviceid();
+                                deviceid = (!deviceid.equals("") ? deviceid : DeviceInfoUtil.getUniquePsuedoID());
+                                device.setDeviceid(deviceid);
+                                device.setTime(time);
+                                LogUtil.debugLog("start connect socketio");
+                                //////////////
+
+                                Map devicemap = DeviceBeanReflect(device);
+                                if(devicemap==null)
+                                    return false;
+                                if (preference.getEchoCustomOption().equals("") == false) {
+                                        Map custompostoption = ExternalInfoUtil.getCustomOption(preference.getEchoCustomOption());
+                                        if (custompostoption != null) {
+
+                                            LogUtil.debugLogWithJava("echo custom option map"+custompostoption.toString());
+                                            if(custompostoption.size()>0)
+                                                devicemap.putAll(custompostoption);
+                                        }
+                                }
+                                echoServerBySocketio(preference.getEchoServer(), gson.toJson(devicemap));
+                                LogUtil.debugLog(gson.toJson(devicemap));
+                                return true;
+
                 }
                 else
                         return false;
@@ -245,6 +262,23 @@ public class NotificationCollectorMonitorService extends Service {
                         this.connectedtime=time;
                 }
 
+        }
+        public Map DeviceBeanReflect(DeviceBean e){
+                Class cls = e.getClass();
+                Field[] fields = cls.getDeclaredFields();
+                Map<String, String> devicebeanmap = new HashMap<String, String>();
+                for(int i=0; i<fields.length; i++){
+                        Field f = fields[i];
+                        f.setAccessible(true);
+                        try {
+                                devicebeanmap.put((String) f.getName(), (String) f.get(e));
+                                //System.out.println("属性名:" + f.getName() + " 属性值:" + f.get(e));
+                        }catch (Exception ee){
+                                //LogUtil.debugLogWithJava(ee.getStackTrace().toString());
+                                return devicebeanmap;
+                        }
+                }
+                return devicebeanmap;
         }
 
         public static class EchoSocket{
