@@ -1,15 +1,20 @@
 package com.weihuagu.receiptnotice;
 
-import android.os.Bundle;
 import android.app.Notification;
+import android.app.PendingIntent;
+import android.os.Bundle;
 import android.service.notification.StatusBarNotification;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+
+import com.weihuagu.receiptnotice.action.ActionStatusBarNotification;
+import com.weihuagu.receiptnotice.action.IDoPost;
+import com.weihuagu.receiptnotice.util.LogUtil;
+import com.weihuagu.receiptnotice.util.NotificationUtil;
+
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
 
-public abstract class NotificationHandle{
+public abstract class NotificationHandle {
         protected String pkgtype;
         protected Notification notification;
         protected Bundle extras;
@@ -19,12 +24,12 @@ public abstract class NotificationHandle{
         protected IDoPost postpush;
         protected ActionStatusBarNotification actionstatusbar;
         public StatusBarNotification sbn;
-        public NotificationHandle(String pkgtype,Notification notification,IDoPost postpush){
-                this.pkgtype=pkgtype;
-                this.notification=notification;
-                this.postpush=postpush;
+        public NotificationHandle(String rawpkgtype, Notification rawnotification, IDoPost rawpostpush){
+                pkgtype=rawpkgtype;
+                notification=rawnotification;
+                postpush=rawpostpush;
 
-                this.extras=notification.extras;
+                extras=notification.extras;
                 // 获取通知标题
                 title = extras.getString(Notification.EXTRA_TITLE, "");
                 // 获取通知内容
@@ -44,46 +49,32 @@ public abstract class NotificationHandle{
                 this.actionstatusbar=actionstatusbar;
         }
         public  abstract void handleNotification();
-        protected  String extractMoney(String content){
-                Pattern pattern = Pattern.compile("(收款|收款￥|向你付款|向您付款|入账)(([1-9]{1}\\d*)|([0]{1}))(\\.(\\d){0,2})?元");
-                Matcher matcher = pattern.matcher(content);
-                if(matcher.find()){
-                        String tmp=matcher.group();
-                        Pattern patternnum = Pattern.compile("(([1-9]{1}\\d*)|([0]{1}))(\\.(\\d){0,2})?");
-                        Matcher matchernum = patternnum.matcher(tmp);
-                        if(matchernum.find())
-                                return matchernum.group();
-                        return null;
-                }else
-                        return null;
 
 
-        }
-        protected boolean predictIsPost(String content){
-                Pattern pattern = Pattern.compile("(收到|收款|向你付款|向您付款|入账)(([1-9]{1}\\d*)|([0]{1}))(\\.(\\d){0,2})?元");
-                Matcher matcher = pattern.matcher(content);
-                if(matcher.find())
-                        return true;
-                else
-                        return false;
-
-        }
 
         protected void removeNotification(){
                 if(actionstatusbar==null|sbn==null)
                         return ;
-                if(predictIsPost(content))
-                        actionstatusbar.removeNotification(sbn);
+                actionstatusbar.removeNotification(sbn);
         }
 
         protected void printNotify(){
                 LogUtil.debugLog("-----------------");
-                LogUtil.debugLog("接受到支付类app消息");
+                LogUtil.debugLog("接受到app消息");
                 LogUtil.debugLog("包名是"+this.pkgtype);
                 NotificationUtil.printNotify(notification);
                 LogUtil.debugLog("**********************");
 
 
+        }
+
+        protected void openNotify(){
+                PendingIntent pendingIntent = notification.contentIntent;
+                try{
+                        pendingIntent.send();
+                }catch(PendingIntent.CanceledException e){
+
+                }
         }
 
 
